+++ packages/calendar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.1](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-calendar@2.0.0...@alfalab/core-components-calendar@2.0.1) (2021-05-31)
+
+**Note:** Version bump only for package @alfalab/core-components-calendar
+
+
+
+
+
 # [2.0.0](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-calendar@1.2.12...@alfalab/core-components-calendar@2.0.0) (2021-05-25)