+++ packages/modal/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.3](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-modal@1.5.2...@alfalab/core-components-modal@1.5.3) (2021-05-31)
+
+**Note:** Version bump only for package @alfalab/core-components-modal
+
+
+
+
+
+## [1.5.2](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-modal@1.5.1...@alfalab/core-components-modal@1.5.2) (2021-05-28)
+
+**Note:** Version bump only for package @alfalab/core-components-modal
+
+
+
+
+
 ## [1.5.1](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-modal@1.5.0...@alfalab/core-components-modal@1.5.1) (2021-05-25)
 
 **Note:** Version bump only for package @alfalab/core-components-modal